windows build complete

+++ Makefile
@@ -1,8 +1,10 @@
 BUILD_DIR := build
 HOST_BUILD_DIR := $(BUILD_DIR)/host
 
+HOST_CXXFLAGS := -std=c++17 -O2 
 HOST_CXX ?= clang++
-HOST_CXXFLAGS := -std=c++17 -O2 -fPIC
+CC ?= clang
+LD ?= ld.lld
 
 DISCORD_SRC_DIR := discord
 DISCORD_SDK_ROOT    := discord/discord_social_sdk
@@ -10,45 +12,61 @@ DISCORD_SDK_INC_DIR := $(DISCORD_SDK_ROOT)/include
 DISCORD_SDK_LIB_DIR := $(DISCORD_SDK_ROOT)/lib/release
 DISCORD_SDK_BIN_DIR := $(DISCORD_SDK_ROOT)/bin/release
 
+###########################################################################
+# Platform-specific Discord SDK settings
+###########################################################################
+
 ifeq ($(OS),Windows_NT)
-    DISCORD_LINK_LIB   := $(DISCORD_SDK_LIB_DIR)/discord_partner_sdk.lib
-    DISCORD_SHARED_BIN := $(DISCORD_SDK_BIN_DIR)/discord_partner_sdk.dll
-    DISCORD_HOST_LIB   := $(HOST_BUILD_DIR)/discord_integration.dll
-    HOST_LDFLAGS       := -shared $(DISCORD_LINK_LIB)
+DISCORD_LINK_LIB    := $(DISCORD_SDK_LIB_DIR)/discord_partner_sdk.lib
+DISCORD_SHARED_BIN  := $(DISCORD_SDK_BIN_DIR)/discord_partner_sdk.dll
+DISCORD_HOST_LIB    := $(HOST_BUILD_DIR)/discord_integration.dll
+DISCORD_RUNTIME_DLL := $(HOST_BUILD_DIR)/discord_partner_sdk.dll
+HOST_LDFLAGS        := -shared "$(DISCORD_LINK_LIB)"
+
+MKDIR := mkdir
+COPY := copy /Y
+fix_path  = $(subst /,\,$1)
 else ifeq ($(shell uname),Darwin)
-    DISCORD_LINK_LIB   := $(DISCORD_SDK_LIB_DIR)/libdiscord_partner_sdk.dylib
-    DISCORD_SHARED_BIN := $(DISCORD_SDK_LIB_DIR)/libdiscord_partner_sdk.dylib
-    DISCORD_HOST_LIB   := $(HOST_BUILD_DIR)/libdiscord_integration.dylib
-    HOST_LDFLAGS       := -shared -Wl,-rpath,'@loader_path' $(DISCORD_LINK_LIB)
+DISCORD_LINK_LIB    := $(DISCORD_SDK_LIB_DIR)/libdiscord_partner_sdk.dylib
+DISCORD_SHARED_BIN  := $(DISCORD_SDK_LIB_DIR)/libdiscord_partner_sdk.dylib
+DISCORD_HOST_LIB    := $(HOST_BUILD_DIR)/libdiscord_integration.dylib
+HOST_LDFLAGS        := -shared -Wl,-rpath,'@loader_path' "$(DISCORD_LINK_LIB)"
+
+MKDIR := mkdir -p
+COPY := cp
+fix_path  = $1
 else
-    DISCORD_LINK_LIB   := $(DISCORD_SDK_LIB_DIR)/libdiscord_partner_sdk.so
-    DISCORD_SHARED_BIN := $(DISCORD_SDK_LIB_DIR)/libdiscord_partner_sdk.so
-    DISCORD_HOST_LIB   := $(HOST_BUILD_DIR)/libdiscord_integration.so
-    HOST_LDFLAGS       := -shared -Wl,-rpath,\$$ORIGIN $(DISCORD_LINK_LIB)
+DISCORD_LINK_LIB    := $(DISCORD_SDK_LIB_DIR)/libdiscord_partner_sdk.so
+DISCORD_SHARED_BIN  := $(DISCORD_SDK_LIB_DIR)/libdiscord_partner_sdk.so
+DISCORD_HOST_LIB    := $(HOST_BUILD_DIR)/libdiscord_integration.so
+HOST_LDFLAGS        := -shared -Wl,-rpath,\$$ORIGIN "$(DISCORD_LINK_LIB)"
+MKDIR := mkdir -p
+COPY := cp
+fix_path  = $1
 endif
 
-ifeq ($(OS),Windows_NT)
-    CC      := "/c/Program Files/LLVM/bin/clang.exe"
-    LD      := "/c/Program Files/LLVM/bin/ld.lld.exe"
-else ifneq ($(shell uname),Darwin)
-    CC      := clang
-    LD      := ld.lld
-else
-    CC      ?= clang
-    LD      ?= ld.lld
-endif
 
-TARGET  := $(BUILD_DIR)/mod.elf
+###########################################################################
+# MIPS compilation & linking flags
+###########################################################################
 
+TARGET := $(BUILD_DIR)/mod.elf
 LDSCRIPT := mod.ld
 ARCHFLAGS := -target mips -mips2 -mabi=32 -O2 -G0 -mno-abicalls -mno-odd-spreg -mno-check-zero-division \
              -fomit-frame-pointer -ffast-math -fno-unsafe-math-optimizations -fno-builtin-memset
-WARNFLAGS := -Wall -Wextra -Wno-incompatible-library-redeclaration -Wno-unused-parameter -Wno-unknown-pragmas -Wno-unused-variable \
-             -Wno-missing-braces -Wno-unsupported-floating-point-opt -Werror=section
+WARNFLAGS := -Wall -Wextra -Wno-incompatible-library-redeclaration -Wno-unused-parameter \
+             -Wno-unknown-pragmas -Wno-unused-variable -Wno-missing-braces \
+             -Wno-unsupported-floating-point-opt -Werror=section
 CFLAGS   := $(ARCHFLAGS) $(WARNFLAGS) -D_LANGUAGE_C -nostdinc -ffunction-sections
 CPPFLAGS := -DMIPS -DF3DEX_GBI_2 -DF3DEX_GBI_PL -DGBI_DOWHILE -I include -I include/dummy_headers \
-            -I mm-decomp/include -I mm-decomp/src -I mm-decomp/extracted/n64-us -idirafter include/libc -idirafter mm-decomp/include/libc
-LDFLAGS  := -nostdlib -T $(LDSCRIPT) -Map $(BUILD_DIR)/mod.map --unresolved-symbols=ignore-all --emit-relocs -e 0 --no-nmagic -gc-sections
+            -I mm-decomp/include -I mm-decomp/src -I mm-decomp/extracted/n64-us \
+            -idirafter include/libc -idirafter mm-decomp/include/libc
+LDFLAGS  := -nostdlib -T $(LDSCRIPT) -Map $(BUILD_DIR)/mod.map \
+            --unresolved-symbols=ignore-all --emit-relocs -e 0 --no-nmagic -gc-sections
+
+###########################################################################
+# File scanning
+###########################################################################
 
 rwildcard = $(foreach d,$(wildcard $(1:=/*)),$(call rwildcard,$d,$2) $(filter $(subst *,%,$2),$d))
 getdirs = $(sort $(dir $(1)))
@@ -65,31 +83,63 @@ ALL_OBJS := $(C_OBJS)
 ALL_DEPS := $(C_DEPS)
 BUILD_DIRS := $(call getdirs,$(ALL_OBJS) $(DISCORD_OBJS))
 
+###########################################################################
+# Build rules
+###########################################################################
+
 all: $(TARGET) $(DISCORD_HOST_LIB)
 
 $(TARGET): $(ALL_OBJS) $(LDSCRIPT) | $(BUILD_DIR)
 	$(LD) $(ALL_OBJS) $(LDFLAGS) -o $@
 
 $(BUILD_DIR):
-	mkdir -p $@
+	$(MKDIR) $(call fix_path,$@)
 
 $(BUILD_DIRS) $(HOST_BUILD_DIR): | $(BUILD_DIR)
-	mkdir -p $@
+	$(MKDIR) $(call fix_path,$@)
 
 $(C_OBJS): $(BUILD_DIR)/%.o : %.c | $(BUILD_DIRS)
 	$(CC) $(CFLAGS) $(CPPFLAGS) $< -MMD -MF $(@:.o=.d) -c -o $@
 
 $(DISCORD_OBJS): $(BUILD_DIR)/%.o : %.cpp | $(BUILD_DIRS)
-	$(HOST_CXX) $(HOST_CXXFLAGS) -I $(DISCORD_SDK_INC_DIR) $(CPPFLAGS) $< -MMD -MF $(@:.o=.d) -c -o $@
+	$(HOST_CXX) $(HOST_CXXFLAGS) -I $(DISCORD_SDK_INC_DIR) $(CPPFLAGS) $< \
+        -MMD -MF $(@:.o=.d) -c -o $@
 
+###########################################################################
+# Copy Discord SDK runtime DLL/SO to host/ for runtime linking
+###########################################################################
+
+ifeq ($(OS),Windows_NT)
+$(DISCORD_RUNTIME_DLL): $(DISCORD_SHARED_BIN) | $(HOST_BUILD_DIR)
+	$(COPY) "$(call fix_path,$<)" "$(call fix_path,$@)"
+else
 $(HOST_BUILD_DIR)/libdiscord_partner_sdk.so: $(DISCORD_SHARED_BIN) | $(HOST_BUILD_DIR)
-	cp $(DISCORD_SHARED_BIN) $@
+	$(COPY) "$(DISCORD_SHARED_BIN)" "$@"
+endif
 
+###########################################################################
+# Build Discord host integration library
+###########################################################################
+
+ifeq ($(OS),Windows_NT)
+$(DISCORD_HOST_LIB): $(DISCORD_OBJS) $(DISCORD_RUNTIME_DLL) | $(HOST_BUILD_DIR)
+	$(HOST_CXX) $(HOST_CXXFLAGS) $(DISCORD_OBJS) $(HOST_LDFLAGS) -o $@
+else
 $(DISCORD_HOST_LIB): $(DISCORD_OBJS) $(HOST_BUILD_DIR)/libdiscord_partner_sdk.so | $(HOST_BUILD_DIR)
 	$(HOST_CXX) $(HOST_CXXFLAGS) $(DISCORD_OBJS) $(HOST_LDFLAGS) -o $@
+endif
+
+###########################################################################
+# Cleaning
+###########################################################################
 
 clean:
+ifeq ($(OS),Windows_NT)
+	if exist $(BUILD_DIR) rmdir /S /Q $(BUILD_DIR)
+else
 	rm -rf $(BUILD_DIR)
+endif
+	
 
 -include $(ALL_DEPS) $(DISCORD_DEPS)
